+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Disable inactive reporters
+- Fix uninitialized constant `Float::NaN`
 
 ## [0.1.3] - 2017-03-12
 ### Fixed